# services/core-api/continuum_api_spec.yaml
openapi: 3.0.3
...
info:
  title: Continuum CRM API
  version: '1.0.0'
  description: |
    An AI-native CRM that captures customer interactions automatically and
    proposes the resulting record changes for human confirmation.

    ## Authentication

    All endpoints except signup, login, provider OAuth callbacks and provider
    webhooks require a bearer token:

        Authorization: Bearer <token>

    Two token types are accepted:

    - **Session JWT** — issued by `/auth/login`, used by the web application.
    - **API token** — prefixed `ctm_`, created in Settings. Subject to the rate
      limit below, and revocable with effect on the very next request.

    ## Rate limiting

    API tokens are limited to 300 requests per minute (FR-API-01). Every
    response carries `X-RateLimit-Limit`, `X-RateLimit-Remaining` and
    `X-RateLimit-Reset`. Requests below the limit are never throttled.

    ## Provenance

    Records that the agent can infer carry a `source` field
    (`HUMAN` | `AGENT_INFERRED` | `IMPORTED` | `ENRICHED`). Deals additionally
    carry `stageSource`, recording whether the current stage was set by a
    person or by a confirmed agent proposal. Consuming clients should surface
    this distinction rather than flattening it.

  contact:
    name: Continuum engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.continuum.example/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /settings/api-tokens:
    get:
      tags: [Settings]
      summary: List API tokens
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { type: object } }
    post:
      tags: [Settings]
      summary: Create an API token
      description: The plaintext token is returned once and never stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  token:
                    type: string
                    description: Shown once. Store it now.
                  prefix: { type: string }
  /settings/api-tokens/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    delete:
      tags: [Settings]
      summary: Revoke an API token
      description: Takes effect on the very next request (FR-API-04).
      responses:
        '204': { description: Revoked }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session JWT, or a `ctm_`-prefixed API token.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...